fix: fixes and refactorings after rebase
