+++ src/course-outline/CourseOutline.scss
@@ -7,7 +7,7 @@
 @import "./empty-placeholder/EmptyPlaceholder";
 @import "./highlights-modal/HighlightsModal";
 @import "./publish-modal/PublishModal";
-@import "./drag-helper/SortableItem";
+@import "../generic/drag-helper/SortableItem";
 @import "./xblock-status/XBlockStatus";
 
 div.row:has(> div > div.highlight) {
