+++ src/course-unit/course-xblock/CourseXBlock.scss
@@ -22,11 +22,6 @@
       font-size: $font-size-sm;
     }
 
-    .pgn__card-header-subtitle-md {
-      margin-top: 0;
-      font-size: 1rem;
-    }
-
     .pgn__card-header-title-md {
       font: 700 1.375rem/1.75rem $font-family-sans-serif;
       color: $black;
