+++ src/course-unit/header-title/messages.js
@@ -26,14 +26,6 @@ const messages = defineMessages({
     defaultMessage: 'Access to some content in this unit is restricted to specific groups of learners.',
     description: 'The label text of some content restriction in this unit',
   },
-  definedVisibilityMessage: {
-    id: 'course-authoring.course-unit.heading.visibility.defined.message',
-    defaultMessage: 'Access to this unit is restricted to: {selectedGroupsLabel}',
-  },
-  commonVisibilityMessage: {
-    id: 'course-authoring.course-unit.heading.visibility.common.message',
-    defaultMessage: 'Access to some content in this unit is restricted to specific groups of learners.',
-  },
 });
 
 export default messages;
